# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Five 网页验证版接口
  version: 1.0.0
  summary: 公开网页、单人后台与后端共同使用的 P0 接口契约
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    本文件是 Five P0 接口字段、状态、错误码和版本保护规则的唯一事实源。
    `contentVersion`、各种 ID 和修订号都是不透明值，调用方不得解析其字符串。
    P0 只有网页端，因此接口不包含 `platform`、`sourcePlatform` 或 `targetPlatform`。
    `channelId` 只用于区分链接分发来源，不得改变日期、档位、正文或图片。
    所有 `/admin/api/v1` 响应（包括错误）必须使用 `Cache-Control: no-store`，防止退出后从缓存读取后台内容或核对凭证。
servers:
  - url: /
    description: 与网页同源的后端
...
security: []
paths:
...
  /admin/api/v1/security-events:
    get:
      operationId: listSecurityEvents
      tags: [Admin security]
      summary: 查询至少保留 365 天的关键登录与安全记录
      description: |
        只返回单向来源指纹和受限浏览器摘要，不返回 IP、密码或会话原文。
      security:
        - AdminSession: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
            maxLength: 256
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: 按发生时间倒序返回安全记录
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityEventPage"
        "400":
          $ref: "#/components/responses/AuthInvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          $ref: "#/components/responses/AdminServiceUnavailable"
  /admin/api/v1/emergency-control:
    get:
      operationId: getEmergencyControlStatus
      tags: [Admin security]
      summary: 查看全局公开内容开关及其修订号
      security:
        - AdminSession: []
      responses:
        "200":
          description: 返回当前公开状态
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/EmergencyControlETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyControlStatus"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          $ref: "#/components/responses/AdminServiceUnavailable"
  /admin/api/v1/emergency-control/stop:
    post:
      operationId: stopPublicAccess
      tags: [Admin security]
      summary: 紧急停止全部公开内容
      description: |
        必须提供精确短语“停止全部公开内容”和原因；会话、Origin、CSRF、并发与幂等保护仍全部生效。
        成功后源站立即 fail closed；已经下载或被外部转发的副本无法召回，正式 CDN 的
        `purge/deny` 联动由 Issue #34 在真实部署环境验收。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/XCsrfToken"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyStopRequest"
      responses:
        "200":
          $ref: "#/components/responses/EmergencyControlSucceeded"
        "400":
          $ref: "#/components/responses/AuthInvalidArgument"
        "401":
          $ref: "#/components/responses/AdminReauthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "409":
          $ref: "#/components/responses/AuthConflict"
        "412":
          $ref: "#/components/responses/RevisionMismatch"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "503":
          $ref: "#/components/responses/AdminServiceUnavailable"
  /admin/api/v1/emergency-control/resume:
    post:
      operationId: resumePublicAccess
      tags: [Admin security]
      summary: 恢复全部公开内容
      description: |
        必须提供精确短语“恢复全部公开内容”和原因；会话、Origin、CSRF、并发与幂等保护仍全部生效。
        恢复只解除全局开关；每个内容版本仍须独立满足已发布、审核和素材安全条件。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/XCsrfToken"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyResumeRequest"
      responses:
        "200":
          $ref: "#/components/responses/EmergencyControlSucceeded"
        "400":
          $ref: "#/components/responses/AuthInvalidArgument"
        "401":
          $ref: "#/components/responses/AdminReauthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "409":
          $ref: "#/components/responses/AuthConflict"
        "412":
          $ref: "#/components/responses/RevisionMismatch"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
        "503":
          $ref: "#/components/responses/AdminServiceUnavailable"
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: five_admin_session
      description: |
        Issue #31 已确认的同源安全会话。生产 Cookie 固定为 `HttpOnly`、`Secure`、
        `SameSite=Strict`、`Path=/admin` 且无 `Domain`；闲置 30 分钟、绝对 12 小时失效。
        所有后台写操作还必须校验可信同源 `Origin` 与会话绑定的 `X-CSRF-Token`。
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: |
        草稿写操作使用响应中的草稿 ETag；生命周期写操作使用内容版本 ETag；
        全局紧急控制使用公开内容开关 ETag。调用方必须原样回传，不能自行拼接。
      schema:
        type: string
        minLength: 3
        maxLength: 256
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 同一业务意图的网络重试必须复用同一个高熵不透明值。
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[-A-Za-z0-9_:.]{16,128}$'
    XCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: |
        当前后台会话绑定的随机 CSRF 令牌。后台写操作还必须由服务端校验可信同源 `Origin`；
        令牌只能保存在运行中页面内存，不能放入 localStorage、URL 或日志。
      schema:
        type: string
        minLength: 32
        maxLength: 256
  headers:
...
    XRequestId:
      description: 当前外部请求的追踪编号；缓存命中时也必须按请求生成。
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    EmergencyControlETag:
      description: 全局公开内容开关当前修订号的强 ETag，例如 `"emergency-control:4"`。
      schema:
        type: string
        example: '"emergency-control:4"'
    NoStoreCacheControl:
      description: 认证、恢复、安全记录和紧急控制响应不得被浏览器或中间层保存。
      schema:
        type: string
        const: no-store
...
  schemas:
...
    ZonedDateTime:
      type: string
      format: date-time
      example: "2026-07-14T23:30:00+08:00"
...
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 128
      description: 不透明标识；调用方不得解析。
...
    SecurityEventAction:
      type: string
      enum:
        - bootstrap_completed
        - login_password
        - logout_current
        - logout_all
        - offline_reset
        - csrf_rejected
        - rate_limited
        - emergency_stop
        - emergency_resume
    SecurityEventOutcome:
      type: string
      enum: [succeeded, rejected]
    SecurityEvent:
      type: object
      additionalProperties: false
      required:
        - eventId
        - occurredAt
        - action
        - outcome
        - requestId
        - sourceFingerprint
        - clientSummary
        - reason
      properties:
        eventId:
          $ref: "#/components/schemas/OpaqueId"
        occurredAt:
          $ref: "#/components/schemas/ZonedDateTime"
        action:
          $ref: "#/components/schemas/SecurityEventAction"
        outcome:
          $ref: "#/components/schemas/SecurityEventOutcome"
        requestId:
          type: string
          minLength: 8
          maxLength: 128
        sourceFingerprint:
          type: string
          minLength: 16
          maxLength: 128
          description: 使用部署密钥生成的单向来源指纹，不是 IP 或可逆地址。
        clientSummary:
          oneOf:
            - type: string
              maxLength: 160
            - type: "null"
          description: 经过截断和字段白名单处理的浏览器摘要，不保存完整 User-Agent。
        reason:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 2000
            - type: "null"
          description: 仅返回紧急控制或离线重置等动作明确记录的审计原因；认证凭据和内部异常不得写入此字段。
    SecurityEventPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SecurityEvent"
        nextCursor:
          oneOf:
            - type: string
              maxLength: 256
            - type: "null"
    EmergencyControlStatus:
      type: object
      additionalProperties: false
      required:
        - publicAccessEnabled
        - revision
        - changedAt
        - reason
        - auditEventId
      properties:
        publicAccessEnabled:
          type: boolean
        revision:
          type: integer
          minimum: 1
        changedAt:
          $ref: "#/components/schemas/ZonedDateTime"
        reason:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 2000
            - type: "null"
        auditEventId:
          oneOf:
            - $ref: "#/components/schemas/OpaqueId"
            - type: "null"
    EmergencyStopRequest:
      type: object
      additionalProperties: false
      required: [confirmationPhrase, reason]
      properties:
        confirmationPhrase:
          type: string
          const: 停止全部公开内容
        reason:
          type: string
          minLength: 1
          maxLength: 2000
    EmergencyResumeRequest:
      type: object
      additionalProperties: false
      required: [confirmationPhrase, reason]
      properties:
        confirmationPhrase:
          type: string
          const: 恢复全部公开内容
        reason:
          type: string
          minLength: 1
          maxLength: 2000
...
    ErrorCode:
      type: string
      enum:
        - INVALID_ARGUMENT
        - INVALID_FORTUNE_DATE
        - AUTHENTICATION_FAILED
        - UNAUTHENTICATED
        - FORBIDDEN
        - CSRF_VALIDATION_FAILED
        - RESOURCE_NOT_FOUND
        - CONTENT_NOT_FOUND
        - LOOK_NOT_FOUND
        - CONTENT_VERSION_CHANGED
        - ACTIVE_CONTENT_VERSION_CHANGED
        - ALGORITHM_FIELD_READ_ONLY
        - DAY_CORRECTION_PAST_DATE
        - DAY_CORRECTION_RELEASE_UNAVAILABLE
        - IDEMPOTENCY_KEY_REUSED
        - EMERGENCY_CONTROL_CONFLICT
        - INVALID_STATE_TRANSITION
        - VERSION_WITHDRAWN
        - HISTORICAL_CONTENT_EXPIRED
        - REVISION_MISMATCH
        - REQUIRED_REVIEW_MISSING
        - MASTER_REVIEW_EVIDENCE_MISSING
        - PUBLISH_PRECHECK_FAILED
        - IMAGE_FILE_INVALID
        - IMAGE_FILE_TOO_LARGE
        - IMAGE_MEDIA_TYPE_UNSUPPORTED
        - IMAGE_REVIEW_INCOMPLETE
        - IMAGE_SET_INVALID
        - IMAGE_WITHDRAWAL_BLOCKED
        - SCHEDULE_TIME_INVALID
        - PRECONDITION_REQUIRED
        - RATE_LIMITED
        - CONTENT_NOT_READY
        - PUBLIC_ACCESS_STOPPED
        - POSTER_GENERATION_UNAVAILABLE
        - FEEDBACK_UNAVAILABLE
        - ANALYTICS_UNAVAILABLE
        - ADMIN_SERVICE_UNAVAILABLE
    Error:
      type: object
      additionalProperties: false
      required: [code, message, retryable, requestId, details]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          minLength: 1
          maxLength: 500
        retryable:
          type: boolean
        requestId:
          type: string
          minLength: 8
          maxLength: 128
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
  responses:
...
    AuthInvalidArgument:
      description: 认证、恢复或高风险确认请求字段无效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    AdminReauthenticationFailed:
      description: 后台会话不存在或已失效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Unauthenticated:
      description: 后台会话不存在或已失效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AdminServiceUnavailable:
      description: 后台凭据库或安全控制服务暂时不可用；调用方应保留当前输入并稍后重试
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          schema:
            type: integer
            minimum: 1
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    CsrfValidationFailed:
      description: 后台写请求的可信同源 `Origin` 或会话绑定 CSRF 令牌校验失败
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    AuthConflict:
      description: 幂等键复用或紧急控制状态发生冲突
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RevisionMismatch:
      description: "`If-Match` 不是资源当前修订号"
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        ETag:
          description: 当前资源 ETag，调用方必须刷新后再操作。
          schema:
            type: string
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            stale:
              $ref: "#/components/examples/RevisionMismatchError"
...
    PreconditionRequired:
      description: 需要并发保护的写操作缺少 `If-Match`
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    EmergencyControlSucceeded:
      description: 全局公开内容开关已在 PostgreSQL 事务中更新并追加安全审计
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
        ETag:
          $ref: "#/components/headers/EmergencyControlETag"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EmergencyControlStatus"
...
  examples:
...
    RevisionMismatchError:
      value:
        error:
          code: REVISION_MISMATCH
          message: 这份内容已被更新，请刷新后重新操作。
          retryable: true
          requestId: req-0002
          details:
            suppliedETag: '"lifecycle:11"'
            currentETag: '"lifecycle:12"'
...